fix: Add HTML resource config for native binaries
This fixes the HTML report generation in the native CLI binaries.
