+++ next/cli-app/src/main/kotlin/dev/vulnlog/cli/parse/reporting/dto/ProjectDataDto.kt
@@ -1,0 +1,7 @@
+package dev.vulnlog.cli.parse.reporting.dto
+
+data class ProjectDataDto(
+    val organization: String,
+    val name: String,
+    val author: String,
+)
