+++ next/cli-app/src/main/kotlin/dev/vulnlog/cli/parse/reporting/dto/ReportDataDto.kt
@@ -1,0 +1,7 @@
+package dev.vulnlog.cli.parse.reporting.dto
+
+data class ReportDataDto(
+    val project: ProjectDataDto,
+    val generatedAt: String,
+    val entries: List<ReportEntryDataDto>,
+)
