+++ next/cli-app/src/main/kotlin/dev/vulnlog/cli/parse/reporting/dto/ReportEntryDataDto.kt
@@ -1,17 +1,5 @@
 package dev.vulnlog.cli.parse.reporting.dto
 
-data class ReportDataDto(
-    val project: ProjectDataDto,
-    val generatedAt: String,
-    val entries: List<ReportEntryDataDto>,
-)
-
-data class ProjectDataDto(
-    val organization: String,
-    val name: String,
-    val author: String,
-)
-
 data class ReportEntryDataDto(
     val primaryId: String,
     val ids: List<String>,
