+++ next/cli-app/src/main/resources/META-INF/native-image/dev.vulnlog/cli/resource-config.json
@@ -1,0 +1,9 @@
+{
+  "resources": {
+    "includes": [
+      {
+        "pattern": "report/.*\\.html$"
+      }
+    ]
+  }
+}
